requirements.txt
asgiref==3.8.1
Django==4.2.18
django-cors-headers==4.4.0
django-phone-field==1.8.1
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
phonenumbers==8.13.53
psycopg2-binary==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
